--- src/main/java/com/demo/qa/test/ui/Util/Environment/Browser.java
@@ -1,5 +1,0 @@
-package com.demo.qa.test.ui.Util.Environment;
-
-public enum Browser {
-    IE, FIREFOX, CHROME, EDGE
-}
