--- src/main/java/com/demo/qa/test/ui/Util/Environment/OS.java
@@ -1,5 +1,0 @@
-package com.demo.qa.test.ui.Util.Environment;
-
-public enum OS {
-    WINDOWS, LINUX, MAC
-}
