+++ src/main/java/com/qa/test/ui/Pages/StartPage.java
@@ -1,0 +1,6 @@
+package com.qa.test.ui.Pages;
+
+public class StartPage extends AbstractPage {
+
+
+}
