+++ src/main/java/com/qa/test/ui/Util/Environment/Browser.java
@@ -1,0 +1,5 @@
+package com.qa.test.ui.Util.Environment;
+
+public enum Browser {
+    IE, FIREFOX, CHROME, EDGE
+}
