+++ src/main/java/com/qa/test/ui/Util/Environment/OS.java
@@ -1,0 +1,5 @@
+package com.qa.test.ui.Util.Environment;
+
+public enum OS {
+    WINDOWS, LINUX, MAC
+}
