+++ src/test/resources/environment.properties
@@ -1,3 +1,2 @@
 browser.name = CHROME
-app.url=https://www.phptravels.net/login
-app.home.url=https://www.phptravels.net/home
+app.url=http://
