+++ src/test/resources/test.properties
@@ -6,8 +6,8 @@ win.gecko.driver=src/main/resources/geckodriver/win64/geckodriver.exe
 mac.gecko.driver=src/main/resources/geckodriver/mac/geckodriver
 win.edge.driver=src/main/resources/edgedriver/
 
-admin.user.name=user@phptravels.com
-admin.user.password=demouser
+admin.user.name=admin
+admin.user.password=password
 test.result.path=target/selenide-reports
 api.url=
 api.authenticate=system.authentication
